migrate snakeyaml
